requirements.txt
Django==5.2.11
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-yasg==1.21.10
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
